refactor: split getters
