+++ src/createRollupEnoughCustomFeeTokenAllowance.ts
@@ -1,7 +1,7 @@
 import { Address, PublicClient, Transport, Chain } from 'viem';
 
 import { fetchAllowance } from './utils/erc20';
-import { getRollupCreatorAddress } from './utils/getters';
+import { getRollupCreatorAddress } from './utils/getRollupCreatorAddress';
 import { createRollupDefaultRetryablesFees } from './constants';
 
 import { Prettify } from './types/utils';
