+++ src/utils/getBlockExplorerUrl.ts
@@ -1,0 +1,5 @@
+import { Chain } from 'viem';
+
+export function getBlockExplorerUrl(chain: Chain | undefined) {
+  return chain?.blockExplorers?.default.url;
+}
