+++ src/utils/getRollupCreatorAddress.ts
@@ -1,0 +1,16 @@
+import { Client, Transport, Chain } from 'viem';
+
+import { rollupCreatorAddress } from '../contracts/RollupCreator';
+import { validateParentChain } from '../types/ParentChain';
+
+export function getRollupCreatorAddress<TChain extends Chain | undefined>(
+  client: Client<Transport, TChain>,
+) {
+  const chainId = validateParentChain(client);
+
+  if (!rollupCreatorAddress[chainId]) {
+    throw new Error(`Parent chain not supported: ${chainId}`);
+  }
+
+  return rollupCreatorAddress[chainId];
+}
