+++ src/utils/getTokenBridgeCreatorAddress.ts
@@ -1,20 +1,7 @@
 import { Client, Transport, Chain } from 'viem';
 
-import { rollupCreatorAddress } from '../contracts/RollupCreator';
 import { tokenBridgeCreatorAddress } from '../contracts/TokenBridgeCreator';
 import { validateParentChain } from '../types/ParentChain';
-
-export function getRollupCreatorAddress<TChain extends Chain | undefined>(
-  client: Client<Transport, TChain>,
-) {
-  const chainId = validateParentChain(client);
-
-  if (!rollupCreatorAddress[chainId]) {
-    throw new Error(`Parent chain not supported: ${chainId}`);
-  }
-
-  return rollupCreatorAddress[chainId];
-}
 
 export function getTokenBridgeCreatorAddress<TChain extends Chain | undefined>(
   client: Client<Transport, TChain>,
@@ -26,8 +13,4 @@ export function getTokenBridgeCreatorAddress<TChain extends Chain | undefined>(
   }
 
   return tokenBridgeCreatorAddress[chainId];
-}
-
-export function getBlockExplorerUrl(chain: Chain) {
-  return chain.blockExplorers?.default.url;
 }
